code/requirements.txt
paho-mqtt
vpython